ibacm: Print correct pkey
[ Upstream commit 4e64a0081f28ab3fdeb3ecc8bfaa2b0aedfb5412 ]

In commit 0161b49e3e1e ("ibacm: Unable to assign EP name for limited
pkey"), a fix was introduced enabling an ipoib interface to be
associated with its corresponding IB port, when a limited pkey was
used.

Said commit missed to change the debug print, such that the real pkey
was printed.

Fixes: 0161b49e3e1e ("ibacm: Unable to assign EP name for limited pkey")
Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>

---

v1 -> v2:
   * Added Ira's r-b
   * Removed Gerrit's Change-Id tag (Håkon)
Signed-off-by: Nicolas Morey-Chaisemartin <nmoreychaisemartin@suse.com>