verbs: Fix pingpong buffer validation
The buffer allocated in the UD test has extra 40 bytes reserved for
GRH for both the client and the server, while the actual payload starts
at offset 40.

Since the buffer validation applies to the payload only, we should take
this offset into account:
As the sender, make sure to fill the payload starting at offset 40 as
all data before that will not be sent on ibv_post_send.
As the receiver, make sure to validate the payload starting at offset 40
as all data before that is not part of the actual payload (GRH/not
valid).

Also, The buffer validation option doesn't require an extra parameter,
remove the extra ':' from all ibv_*_pingpong examples.

Fixes: 099c5aa50bc8 ("libibverb/examples: Add command line option to enable buffer validation")
Cc: Yuval Shaia <yuval.shaia@oracle.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>