pyverbs: Introduce PD class
This patch introduces PD object which represents ibv_pd.

This commit also introduces the usage of weak references.
In Python, a weak reference is a reference to an object not strong
enough to keep it from being garbage-collected.
In Pyverbs, it is used to allow an object to keep track of the objects
created using it, so if an object is being destroyed (e.g. explicitly
by the user), it can iterate over the dependent objects and destroy
the underlying C objects.
E.g.: A user opens a device and creates a PD. If the Context object
is being destroyed before the PD object, the kernel will return
EBUSY. To avoid that, the Context object will hold a weakref to the
PD and destroy the C PD before closing itself.
This renders the Python PD object useless, but it is a user's choice
that we are not blocking.

In order to provide a clean teardown, Pyverbs Context is also a
context manager, which means itshouldbe used within a 'with' block.
This way, when the block is over, teardown will be executed properly.
The same applies for all other relevant pyverbs' classes.
A PyverbsCM class was added as an alternative base objects to inherit
from. It inherits from PyverbsObject and adds the __enter__ and
__exit__ functions needed for a Python context manager.

Signed-off-by: Noa Osherovich <noaos@mellanox.com>