verbs: Demonstrate the usage of new post send API
Expose a new flag in ibv_rc_pingpong command line so send WRs will be
posted by the new post send method as introduced previously by
libibverbs. It can be used as a simple example for this API usage, as a
sanity for this API functionality or for any other goal.

We use this opportunity to fix the man page with device memory option
description which was missed.

Signed-off-by: Guy Levi <guyle@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>