libhns: Bugfix for filtering zero length sge
[ Upstream commit 1c2def4817f31c15342669144ef6bb894bd9616d ]

When user posts a wqe with n sges, the driver needs to
determine to if each sge is valid and filter the invalid
sge, fill the number of valid sges into sq wqe. Hip08
hardware allocates the correct memory location for each
valid sge based on the number of valid sge. For exmaple,
when posts a wqe with 3 sges. if 3 sge is valid, the first
and second sges will be filled into the wqe and the last
sge will fill the indpendent memory location which the
hardware allocated additionally. However, it will happen
error if the first and the second sges is invalid.
Because the driver will filter the first two sge and fill
the last sge into hardware. But the valid sge will be
stored in the extended memory and the correct way
should be stored in wqe.

The other situation example as follows:
Posting five sge and have two invalid sge.
 ________________________________________
|___0____|__1____|___0____|__1___|___1__|

the 0 express the sge is invalid, the 1 express the sge is valid.
Based on the above situation. the hardware will store the second
and the fourth sge into sq wqe. The last sge will be stored in
the extend memory location. the number of valid sge is 3.

Fixes: 29ef2625c4e1 ("libhns: Filter for zero length of sge in hip08 userspace")
Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: Nicolas Morey-Chaisemartin <nmoreychaisemartin@suse.com>