mlx5: Support scatter to CQE over DCT QP
Scatter to CQE which is a performance feature has never been enabled on
DCT QP.

A new capability which allows to enable the feature on DCT QP is
reported in query device related verb. Correspondingly, this patch
enables the feature functionality so it is aligned with other QPs and
the man page. It will be controlled by the legacy environment variable
and can be override by the DV creation flags.

Fixes: b9967a9d722a ("mlx5: Create DC transport QPs")
Signed-off-by: Guy Levi <guyle@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>