pyverbs: Introducing QP class
Allow creation, modification, querying and post recv and send
operations on QPs of types RC, UC, UD and Raw Packet.

Creation process allows user to get the QP in a more advanced state
than INIT; by providing a QPAttr object (in addition to the mandatory
QPInitAttr object), the user tells pyverbs to use the provided values
and transition the QP as far as possible towards RTS. For connected
QPs it means to INIT state, for UD / Raw Packet QPs it means to RTS.
By default, the QPAttr's value will be None and the QP will be
returned in RESET state.

The QP class offers to_init/rtr/rts methods for quick and easy state
transitions - they only require the user to provide a QPAttr object
and the comp masks are being set according to the bits required for
each QP type / transition.

Similarly to querying a device, querying the QP doesn't require
providing the QPAttr / QPInitAttr objects; they are created by
pyverbs and are returned to the user as a tuple.

This patch also adds weakref support to CQ/CQEX to keep pyverbs'
reference scheme.

Signed-off-by: Noa Osherovich <noaos@mellanox.com>