vmw_pvrdma: Use resource ids from physical device if available
This change allows user-space to use physical resource numbers if
they are passed up from the driver. Doing so allows communication with
physical non-ESX endpoints (such as a bare-metal Linux machine or a
SR-IOV-enabled VM).

This is accomplished by separating the concept of the QP number from
the QP handle. Previously, the two were the same, as the QP number
was exposed to the guest and also used to reference  a virtual QP in
the device backend. With physical resource numbers exposed, the QP
number given to the guest is the QP number assigned to the physical
HCA's QP, while the QP handle is still the internal handle used to
reference a virtual QP. Regardless of whether the device is exposing
physical ids, the driver will still try to pick up the QP handle from
the backend if possible.

The MR keys exposed to the guest when the physical resource ids feature
is turned on are likewise now the MR keys created by the physical HCA,
instead of virtual MR keys.

The ABI has been updated to allow the return of the QP handle to the
guest library. The ABI version has been bumped up because of this
non-compatible change.

Reviewed-by: Jorgen Hansen <jhansen@vmware.com>
Signed-off-by: Adit Ranadive <aditr@vmware.com>
Signed-off-by: Bryan Tan <bryantan@vmware.com>