verbs: Remove verbs_device sysfs_path
This is mainly used during the discovery loop to form paths relative to
the uverbsX class directory. To avoid races/etc with the kernel this is
better done using fd relative *at functions.

Preserve the ABI of the struct ibv_device by regenerating the class prefix
when the ibv_device is created.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>