RDMA/hns: Bugfix for identify the last srq sge
When post multiple sges with a srq, the hardware will
identify the last sge according to the Specific identifier
with lkey. when the user issue sge with every wr less
than the max sge that the user configured, the hardware
will stop getting sge according to the sepcific lkey.
However, it will always end with the first sge in the
current post srq recv interface implementation.

Fixes: 4332ae4947bb ("libhns: Add the verb for posting srqwqe")
Signed-off-by: Lijun Ou <oulijun@huawei.com>