Enhance build system to allow separate installation
This is achieved by semi-automatically assigning a component name based
on the current source dir.
For static libs, a seperate build is not sensibly possible due to how
they are generated. But it's at least made sure all the dependencies are
cleanly built, and ultimately only the requested lib is built.

With this, one can for example do "make install-libibverbs" to install
just libibverbs and its related files. Man pages and tests are
intentionally assigned a separate target, like install-libibverbs-man.

Signed-off-by: BtbN <btbn@btbn.de>