mlx5: Add support for bulk flow counters steering actions
Flow counters can be allocated/freed one by one, or as a consecutive
range of counters - bulk.
When the counters are allocated as a bulk, each of the counters can be
assigned to a separate steering action. Members of this counter range
are addressed by a base counter ID and an offset.

Add the above support to work over root/kernel flows when the DR flow
counters API was used.

Signed-off-by: Yevgeny Kliteynik <kliteyn@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>