kernel-boot: Instrument rename utility to allow sane bug reports
Generate the log messages, which will be delivered by systemd-journal to
easy debug and bug reports filled by users.

In order to see debug messages in systemd log, there is a need to
add "-v" as a parameter to rdma_renam utility in udev rule.

The inability to generate debug mode for specific udev rules pushed us
to implement custom pr_dbg() macros with LOG_ERR inside and not LOG_DEBUG.

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>