mlx5: DR, Remove table level check
To prevent loops a check was done to fail on connecting tables to lower
levels.
Since SWS is allowed only on trusted systems, with privileged users,
current decision is to leave loop prevention for the privileged user to
deal with, in this case a user can build more complex table trees and
reuse the same table, for example using the same 5tuple match table for
inner and outer, by decapsulating and jumping to the same table.

Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>
Signed-off-by: Yishai Hadas <yishaih@nvidia.com>