debian: Add Pre-Depends on ${misc:Pre-Depends}
When using debhelper 12, lintian will complains about
skip-systemd-native-flag-missing-pre-depends:

This package uses the --skip-systemd-native invoke-rc.d flag in the
specified maintainer script but does not specify a Pre-Depends
dependency on a recent version of init-system-helpers.

This flag is useful for maintainer scripts that want to defer systemd
actions to deb-systemd-invoke(1p). However, it was only added in
init-system-helpers version 1.58.

Please add Pre-Depends: ${misc:Pre-Depends} to your debian/control
file.

Refer to the invoke-rc.d(8) manual page and the deb-systemd-invoke(1p)
manual page for details.

Signed-off-by: Benjamin Drung <benjamin.drung@cloud.ionos.com>