build/azp: Use a version number for the docker images
This way when we make a stable branch the CI container images will remain
unchanged for that branch instead of tracking the master branch.

If incompatible changes are made during a release then a suffix should be
appended.

The version number for the image is taken from the
buildlib/azure-pipelines.yml.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>