build/azp: Have Azure Pipelines create releases when tags are made
[ Upstream commit 936cf8e17cf30dd37ed9b9bb91576603df9184ae ]

Build a draft release and push the source .tar.gz to GitHub's releases
page.

For some reason this relies on an Azure Pipeline's user-specific service
connection instead of the existing app installation:

https://github.com/MicrosoftDocs/vsts-docs/issues/4260

This is done in a dedicated YAML file as it runs with a separate
configuration on AZP that authorizes it to use the write-access service
connections.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>