libhns: Not process return value of flushing cqe
[ Upstream commit 482eb44bc2960ec628de91ae32d1c9571ed686e0 ]

When a firmware reset occurred because of some errors, modify qp to error
in post_send() will be failed. The bad_wr will be returned to the user,
then the user won't wait for the corresponding wc. But as sq db has been
updated, the driver will return a software wc to the user, which may cause
unexpected behaviours.

Because the result of modifying qp is meaningless in this scenario, codes
to handle the return value are removed.

Signed-off-by: Yangyang Li <liyangyang20@huawei.com>
Signed-off-by: Jiaran Zhang <zhangjiaran@huawei.com>
Signed-off-by: Weihang Li <liweihang@huawei.com>
Signed-off-by: Nicolas Morey-Chaisemartin <nmoreychaisemartin@suse.com>