efa: WQE format cleanups
This patch introduces a few cleanups to the WQE format.

Misc comments were reworded.

The meta union is "inlined" to what was previously called "common" meta
descriptor, which is now the meta descriptor (no binary interface
changes).

The EFA lkey is 24 bits long, as already seen in the RX descriptor,
align the TX WQE as well. The rest of the bits are turned into reserved.

Change the length field to 32 bits by taking over unused (reserved) bits.
Note that the length for send operations is limited by max_msg_size
which is less than 16 bits, hence no backwards compatibility issues.

Signed-off-by: Gal Pressman <galpress@amazon.com>