efa: Fix unmasked QP table access
The QP table size is set according to the maximum number of QPs.
However, it is not guaranteed that the QP numbers must be assigned in
the range of [0, max_qp). The device guarantees that the QP number LSBs
(according to the log of max_qp) are unique, hence access to the QP
table must be accompanied with a proper masking.

Signed-off-by: Gal Pressman <galpress@amazon.com>