debian: Specify Build-Depends-Package for libraries
The symbol files for libibmad.so.5 and libibnetdisc.so.5 do not contain
a Build-Depends-Package meta-information field.

This field specifies the name of the -dev package associated to the
library and is used by dpkg-shlibdeps(1) to make sure that the
dependency generated is at least as strict as the corresponding build
dependency.

This is useful as allows packages to not hardcode this information
multiple times.

Refer to the deb-symbols(5) manual page and the dpkg-shlibdeps(1) manual
page for details.

Signed-off-by: Benjamin Drung <benjamin.drung@cloud.ionos.com>