mlx5: Limit STE hash table enlarge based on bytemask
When an ste hash table has too many collision we enlarge it
to a bigger hash table (rehash). Rehashing collision improvement
depends on the bytemask value. The more 1 bits we have in bytemask
means better spreading in the table.

This patch will limit table rehash to reduce memory and improve
the performance, based on bytemask.

Fixes: a91e8c2bbedf ('mlx5: Expose steering rule functionality')
Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>