cbuild: Check if file exists before creating a symbolic link
In case the tar already exists with the same name there's no need to
create a symbolic link for it. This resolves an issue where the symlink
call would fail due to FileExistsError exception.

This is observed when packaging AmazonLinux1 that uses centos6 spec file
which uses the name rdma-core.tgz.

Signed-off-by: Gal Pressman <galpress@amazon.com>