tests: Add a test for parent domain
The test checks parent domain creation without custom allocators, with
custom allocators that return USE_DEFAULT (indicates the driver to use
the default allocator) and custom allocators with local memory
allocation.
In addition, in each case a QP and SRQ are created in order to verify
the usage of the allocators.

Signed-off-by: Edward Srouji <edwards@mellanox.com>
Reviewed-by: Noa Osherovich <noaos@mellanox.com>