tests: Make multiprocessing available to more than a single test
set_start_method(), which can only be called once, was used in the
TestCase's setUp() method.
If more than one test is executed, setUp() is called more than once and
an error occurs.
Replace set_start_method() with Python's multiprocessing's context and
move the call to fork to the test itself.

Signed-off-by: Maxim Chicherin <maximc@mellanox.com>
Reviewed-by: Daria Velikovsky <daria@mellanox.com>
Reviewed-by: Noa Osherovich <noaos@mellanox.com>