mlx5: Add support for meter in SW steering
Add support to create a flow meter action based on the flow meter
parameters. The flow meter parameters are according to the device
specification.

A flow meter can be created either in active or not active state, the
state could be changed by using the modify API. In addition to the
meter parameters, the user should provide the next flow table in the
matching process.

Signed-off-by: Maor Gottlieb <maorg@mellanox.com>
Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>