verbs: Relaxed ordering memory regions
Add a flag to allow creation of relaxed ordering memory regions.
Access through such MRs can improve performance by allowing the
system to reorder certain accesses.

As relaxed ordering is an optimization, drivers that do not support
it can simply ignore it.

An optional MR access bit range is defined based on the kernel
matching part and its first entry will be IBV_ACCESS_RELAXED_ORDERING.

In case an application uses one of the bits from the optional range
the library will mask it out in case the 'MR optional mode' is not
supported by the kernel.

Signed-off-by: Michael Guralnik <michaelgur@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>