vmw_pvrdma: Use QP handle when attempting to flush CQEs
[ Upstream commit 68c623814d8d62c6b09c5d5c6390bc2286631e64 ]

After the change to distinguish QP numbers and QP handles, the
cqe->qp field contains the QP handle so as to make lookup for
QPs simple. Flushing CQEs upon destroy of QP needs to be done
by comparing the cqe->qp field with the QP handle, not the QP
number. This change fixes this issue.

Fixes: 3dbaeba640d7  ("vmw_pvrdma: Use resource ids from physical device
if available")
Reviewed-by: Jorgen Hansen <jhansen@vmware.com>
Reviewed-by: Rajesh Jalisatgi <rjalisatgi@vmware.com>
Signed-off-by: Bryan Tan <bryantan@vmware.com>
Signed-off-by: Nicolas Morey-Chaisemartin <nmoreychaisemartin@suse.com>