mlx5: Remove uneeded reserved fields from dr internal match structs
The reserved fields are not needed since we set each and every value
from the PRM big endian to the local endian fields using an endian
agnostic macro. In the past this was used to keep the PRM and local
struct in the same size for copy and endian conversion.

Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>