efa: Fail context allocation in case create AH verb doesn't have udata
The EFA provider cannot work correctly when there's no udata in
create_ah verb, as we must get the AH number from the kernel.

Without this patch, the provider assumes the AH number is zero for all
AHs, which results in unexpected behavior.
On context allocation, identify kernels that don't support udata and
exit.

Signed-off-by: Gal Pressman <galpress@amazon.com>