libqedr: Fix user context allocation forward compatibility
[ Upstream commit 9bb0fcf83c305ab88811d3de068977700d3eb78b ]

The user context alloc request structure introduced a new field
to indicate doorbell recovery is supported. This field was added so
that additional features could be added in the future by setting a
capability flag. However, the field wasn't zeroed, and was initialized
using "|=" instead of "=" leading to garbage in the other bits.
For forward compatability, we need to make sure all other bits are
zero.

Cc: stable@linux-rdma.org # v27 v28
Fixes: d9b2ba480af5 ("libqedr: Add support for Doorbell Overflow Recovery")
Signed-off-by: Ariel Elior <ariel.elior@marvell.com>
Signed-off-by: Michal Kalderon <michal.kalderon@marvell.com>
Signed-off-by: Nicolas Morey-Chaisemartin <nmoreychaisemartin@suse.com>