mlx5: Set flow_label and RoCEv2 UDP source port for datagram QP
The flow_label and UDP source port fields are used to create entropy for
network routers (ECMP), load balancers and 802.3ad link aggregation
switching that are not aware of RoCE IB headers. This patch set them
properly for UD packets.

When create an AH, the UDP source port is calculated and set base on
flow_label if it's not zero:
    #define IB_ROCE_UDP_ENCAP_VALID_PORT_MIN 0xC000
    u32 fl_low  = fl & 0x03FFF;
    u32 fl_high = fl & 0xFC000;
    u16 udp_sport = fl_low XOR (fl_high >> 14);
    RoCE.UDP.src_port = udp_sport OR IB_ROCE_UDP_ENCAP_VALID_PORT

otherwise a random value is used.

Ref.: https://www.spinics.net/lists/linux-rdma/msg87626.html

Signed-off-by: Mark Zhang <markz@mellanox.com>
Reviewed-by: Maor Gottlieb <maorg@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>