rsockets: Do not overwrite connection failure error code
In rs_do_connect, if the rsocket state is not connected, we set
errno = EINVAL and ret = -1.  Later in the code, we modify the
rsocket state to an error state and save errno with the rsocket.
Apps can obtain the error data through getsockopt.

This can result in overwriting real connection error data that
may have been saved with the rsocket.  The JDK catches this error
in a negative test case.  It reads the EINVAL error code from the
socket, when it expects ECONNREFUSED.

Update the code so that we only save the error with the socket if
it came from a connection error.  In other cases, simply set errno
and return a failure.  We add explicit state checks for the other
non-connected rsocket cases.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>