mlx5: Fix postsend actions write length
[ Upstream commit a35be3d87cf1846afc49c54a8b9efab52366cc3b ]

Adapt the send info write length to the dynamic number of actions.

Before changing the actions number to be dynamic the length matched to
the chunk capacity.
Post moving to support the dynamic mode the real length should be used
to prevent copying from smaller data buffer a larger chunk capacity
which might lead to a segmentation fault.

Fixes: 9d1874342ff0 ("mlx5: Extended modify header to support more than 8 modify actions")
Signed-off-by: Hamdan Igbaria <hamdani@mellanox.com>
Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Nicolas Morey-Chaisemartin <nmoreychaisemartin@suse.com>