d/control: Don't build-depend on pandoc on i386
Ubuntu i386 no more has pandoc and would stall on build dependencies.
This was done as an emergency unblock late in the 20.04 cycle,
but it turned out that i386 still gets all the man pages that pandoc
would convert properly installed.
See amd64 [1] with pandoc vs i386 (without pandoc) [2] and as an example
the man page ibv_resize_cq.3.

[1]: https://launchpadlibrarian.net/466565153/buildlog_ubuntu-focal-amd64.rdma-core_28.0-1ubuntu1_BUILDING.txt.gz
[2]: https://launchpadlibrarian.net/466565266/buildlog_ubuntu-focal-i386.rdma-core_28.0-1ubuntu1_BUILDING.txt.gz

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>